evaluation_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain==0.2.6
langchain-anthropic==0.1.19
anthropic==0.30.0
pydantic==2.7.4
redis==5.0.7
python-dotenv==1.0.1
httpx==0.27.0
